pyproject.toml
[project]
name = "mcp-lang-graph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.4.400",
    "faiss-cpu>=1.10.0",
    "httpx>=0.28.1",
    "huggingface-hub>=0.30.1",
    "langchain>=0.3.22",
    "langchain-community>=0.2.17",
    "langchain-fireworks>=0.1.7",
    "langchain-huggingface>=0.1.2",
    "langchain-mcp-adapters>=0.0.3",
    "langchain-openai>=0.3.11",
    "langgraph>=0.3.21",
    "langgraph-cli[inmem]>=0.1.84",
    "langsmith>=0.3.24",
    "mcp[cli]>=1.6.0",
    "pymupdf>=1.25.5",
    "pymysql>=1.1.1",
    "python-dotenv>=1.0.1",
    "sentence-transformers>=4.0.2",
    "tavily-python>=0.4.0",
]

